pyproject.toml
[tool.poetry]
name = "flask-sqlorm"
version = "0.1.0"
description = "Flask integration for sqlorm"
authors = ["Maxime Bouroumeau-Fuseau <maxime.bouroumeau@gmail.com>"]
readme = "README.md"
packages = [{include = "flask_sqlorm.py"}]

[tool.poetry.dependencies]
python = "^3.8"
sqlorm-py = "^0.1.2"
flask = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"

[tool.poetry.group.postgresql.dependencies]
psycopg = {extras = ["binary"], version = "^3.1.18"}

[tool.poetry.group.mysql.dependencies]
mysql-connector-python = "^8.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"